requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.2
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
requests>=2.28.0
supabase>=2.15.0
pytest>=7.3.1
gunicorn>=20.1.0
